STEP 3 - EDA Analysis/requirements.txt
jupyterlab
pandas
sqlalchemy
psycopg2-binary
matplotlib
seaborn
plotly
